+++ clx_invoice_policy/models/sale_subscription.py
@@ -12,7 +12,6 @@ from odoo import fields, models, api, _
 from calendar import monthrange
 
 
-
 class SaleSubscription(models.Model):
     _inherit = "sale.subscription"
 
@@ -69,7 +68,8 @@ class SaleSubscriptionLine(models.Model):
     cancel_invoice_end_date = fields.Date('Cancel End Date')
     account_id = fields.Many2one('account.move', string="Invoice")
     is_prorate = fields.Boolean(related="so_line_id.is_prorate", string="Is Prorate?", readonly=False)
-    prorate_amount = fields.Float(related="so_line_id.prorate_amount", string="Prorate Amount", readonly=False)
+    prorate_amount = fields.Float(related="so_line_id.prorate_amount", string="Prorate Start Amount", readonly=False)
+    prorate_end_amount = fields.Float(string="Prorate End Amount")
 
     def _creation_next_budgets(self):
         print("CRON CRON CRON")
@@ -258,6 +258,14 @@ class SaleSubscriptionLine(models.Model):
                     'management_fees': new_management_price,
                     'wholesale': new_price - new_management_price,
                 })
+            if self.end_date.month == self.invoice_end_date.month:
+                new_price = self.prorate_end_amount
+                new_management_price = line.management_price / 2
+                res.update({
+                    'price_unit': new_price,
+                    'management_fees': new_management_price,
+                    'wholesale': new_price - new_management_price,
+                })
         if line.display_type:
             res['account_id'] = False
         if self._context.get('advance', False):
@@ -312,7 +320,8 @@ class SaleSubscriptionLine(models.Model):
                             'invoice_end_date': self.end_date
                         }
                     )
-                elif self.invoice_start_date and self.invoice_start_date > self.end_date and not self._context.get('generate_invoice_date_range', False):
+                elif self.invoice_start_date and self.invoice_start_date > self.end_date and not self._context.get(
+                        'generate_invoice_date_range', False):
                     self.with_context(skip=True).write(
                         {
                             'invoice_end_date': False,
@@ -350,8 +359,7 @@ class SaleSubscriptionLine(models.Model):
                     'name': period_msg,
                 })
                 if self.is_prorate:
-                    if (self.end_date and self._context.get('end_date').month == end_date.month) or (
-                            self.start_date and self._context.get('start_date').month == start_date.month):
+                    if (self.start_date and self._context.get('start_date').month == start_date.month):
                         new_price = self.prorate_amount
                         new_management_price = line.management_price / 2
                         res.update({
@@ -359,4 +367,13 @@ class SaleSubscriptionLine(models.Model):
                             'management_fees': new_management_price,
                             'wholesale': new_price - new_management_price,
                         })
+                    if (self.end_date and self._context.get('end_date').month == end_date.month):
+                        new_price = self.prorate_end_amount
+                        new_management_price = line.management_price / 2
+                        res.update({
+                            'price_unit': new_price,
+                            'management_fees': new_management_price,
+                            'wholesale': new_price - new_management_price,
+                        })
+
         return res